encoder: fixed issue #3
The problems described in issue #3 are because of the padding added when the
image size can't be divided by 8, i.e., some macroblocks will not be complete
or 8x8.  This new padded size is used in quantization and DCT, but until now it
wasn't used when reading YUV data. Using calloc() we now memset the (slightly
larger) allocated memory, making it a bit more initialized.